+++ environment.yml
@@ -23,6 +23,11 @@ dependencies:
   - gitpython >=2.1.0
   - invoke
   - cryptography >=47.0.0
+  # m2crypto, pyasn1 and pyasn1-modules are only needed for the legacy
+  # M2Crypto based implementation (DIRAC_USE_M2CRYPTO=Yes)
+  - m2crypto >=0.38.0
+  - pyasn1 >0.4.1
+  - pyasn1-modules
   - matplotlib
   - numpy
   - paramiko
@@ -87,7 +92,10 @@ dependencies:
   - uritemplate
   # - readline >=6.2.4 in the standard library
   - simplejson >=3.8.1
-  - tornado >=5.1.1,<6.0.0
+  # tornado is installed with pip below: the DIRACGrid fork is needed for the
+  # legacy M2Crypto based implementation (DIRAC_USE_M2CRYPTO=Yes) and behaves
+  # as the standard tornado >=5.1.1,<6.0.0 otherwise
+  #- tornado >=5.1.1,<6.0.0
   - typing >=3.6.6
   - rucio-clients >=34.4.2
   # For mypy
@@ -103,6 +111,13 @@ dependencies:
     - Authlib >=1.0.0
     - dominate
     - pyjwt
+    # The following two packages are only needed for the legacy M2Crypto
+    # based implementation (DIRAC_USE_M2CRYPTO=Yes)
+    # This is a fork of tornado with a patch to allow for configurable iostream
+    - git+https://github.com/DIRACGrid/tornado.git@iostreamConfigurable
+    # This is an extension of Tornado to use M2Crypto
+    # It should eventually be part of DIRACGrid
+    - git+https://github.com/DIRACGrid/tornado_m2crypto
     - -e .[server]
     - -e ./dirac-common/
     # Add diracdoctools
