+++ setup.cfg
@@ -101,6 +101,12 @@ testing =
     pytest-mock
     pytest-rerunfailures
     pycodestyle
+# Only needed for the legacy M2Crypto based implementation (DIRAC_USE_M2CRYPTO=Yes)
+m2crypto =
+    M2Crypto >=0.36
+    pyasn1
+    pyasn1-modules
+    tornado-m2crypto
 
 [options.entry_points]
 dirac =
