+++ src/DIRAC/Core/Security/pyca/X509CRL.py
@@ -6,7 +6,7 @@ import datetime
 from cryptography import x509
 
 from DIRAC import S_OK, S_ERROR
-from DIRAC.Core.Security import asn1_utils
+from DIRAC.Core.Security.pyca import asn1_utils
 from DIRAC.Core.Utilities import DErrno
 from DIRAC.Core.Utilities.File import secureOpenForWrite
