+++ src/DIRAC/Core/Security/pyca/__init__.py
@@ -1,0 +1,1 @@
+"""X509 implementation based on pyca/cryptography and the standard library ssl module"""
